+++ gfa-anagrafiche/src/main/java/gov/mef/gfa/anagrafiche/service/BeneficiarioService.java
@@ -1,9 +1,11 @@
 package gov.mef.gfa.anagrafiche.service;
 
+import java.math.BigDecimal;
+
 import gov.mef.gfa.anagrafiche.exception.ServiceException;
 import gov.mef.gfa.common.bean.anagrafica.BeneficiarioRes;
 import reactor.core.publisher.Mono;
 
 public interface BeneficiarioService {
-	public Mono<BeneficiarioRes> getBeneficiarioById(long id) throws ServiceException;
+	public Mono<BeneficiarioRes> getBeneficiarioById(BigDecimal id) throws ServiceException;
 }
